Extract economy country form to own component
